+++ DuckDB.NET.Data/Internal/ScalarFunctionInfo.cs
@@ -6,11 +6,11 @@ using DuckDB.NET.Native;
 
 namespace DuckDB.NET.Data.Internal;
 
-class ScalarFunctionInfo(IReadOnlyCollection<DuckDBLogicalType> parameterTypes, DuckDBLogicalType returnType, Action<VectorDataReaderBase[], VectorDataWriterBase, ulong> action) : IDisposable
+class ScalarFunctionInfo(IReadOnlyCollection<DuckDBLogicalType> parameterTypes, DuckDBLogicalType returnType, Action<VectorDataReaderBase[], VectorDataWriterBase, int> action) : IDisposable
 {
     public DuckDBLogicalType ReturnType { get; private set; } = returnType;
     public IReadOnlyCollection<DuckDBLogicalType> ParameterTypes { get; private set; } = parameterTypes;
-    public Action<VectorDataReaderBase[], VectorDataWriterBase, ulong> Action { get; private set; } = action;
+    public Action<VectorDataReaderBase[], VectorDataWriterBase, int> Action { get; private set; } = action;
 
     public void Dispose()
     {
